Show Modal and other problems fixed
Issue number 1: Cannot set property 'innerHTML' of null at showModal (scripts.js:81) 

Issue number 2: let modalContainer = document.querySelector(‘#modalContainer’) with this problem the querying was aimed to a wrong id which is modal-container. I changed that and now they are matching. 

Issue number 3: Typo in CloseButtonElement, it was ClsoseButtonElement. Therefore the console was displaying ReferenceError: clsoseButtonElement is not defined at showModal. 

Issue number 4: Uncaught (in promise) ReferenceError: pokemon is not defined at showModal. With this problem the parameter wasn't matching. Instead of writing poke_mon, I wrote pokemon that's why the console was showing ReferenceError: pokemon is not defined at showModal.

Issue number 5: After fixing those errors, the modal was showing, but once the buttons were clicked it was displaying undefined because the arguments which was passing trough ShowModal was not poke_mon, but showModal(poke_mon.name, poke_mon.height);